+++ packages/variance/src/core.ts
@@ -12,6 +12,7 @@ import {
 } from './types/config';
 import { AbstractTheme, CSSObject, ThemeProps } from './types/props';
 import { AllUnionKeys, KeyFromUnion } from './types/utils';
+import { getStaticCss } from './utils/getStaticProperties';
 import { orderPropNames } from './utils/propNames';
 import {
   arrayParser,
@@ -50,7 +51,7 @@ export const variance = {
                 );
               // handle any props configured with the responsive notation
               case 'object':
-                // If it is an array the order of values is smallest to largest: [base, xs, ...]
+                // If it is an array the order of values is smallest to largest: [_, xs, ...]
                 if (isMediaArray(value)) {
                   return merge(
                     styles,
@@ -136,15 +137,40 @@ export const variance = {
         P extends Parser<T, TransformerMap<T, Config>>
       >(config: Config): CSS<T, P> {
         const parser = this.create(config);
-
+        const filteredProps: string[] = [...parser.propNames];
         return (cssProps) => {
           let cache: CSSObject;
+          const allKeys = Object.keys(cssProps);
+
+          /** Any key of the CSSProps that is not a System Prop or a Static CSS Property is treated as a nested selector */
+          const selectors = allKeys.filter(
+            (key) => !filteredProps.includes(key) && isObject(cssProps[key])
+          );
+
+          /** Static CSS Properties get extracted if they match neither syntax */
+          const staticCss = getStaticCss(cssProps, [
+            ...selectors,
+            ...filteredProps,
+          ]);
 
           return ({ theme }) => {
             if (cache) return cache;
             const css = parser({ ...cssProps, theme } as any);
-            cache = css;
+            selectors.forEach((selector) => {
+              const selectorConfig = cssProps[selector];
+              if (isObject(selectorConfig)) {
+                css[selector] = {
+                  ...getStaticCss(selectorConfig, filteredProps),
+                  ...parser(Object.assign(selectorConfig, { theme }) as any),
+                };
+              }
+            });
 
+            /** Merge the static and generated css and save it to the cache */
+            cache = {
+              ...staticCss,
+              ...css,
+            };
             return cache;
           };
         };
@@ -168,7 +194,7 @@ export const variance = {
           Object.keys(options.variants).forEach((key) => {
             const variantKey = key as Keys;
             const cssProps = options.variants[variantKey];
-            variantFns[variantKey] = css(cssProps);
+            variantFns[variantKey] = css(cssProps as any);
           });
 
           return ({ [prop]: selected = defaultVariant, ...props }) => {